+++ benches/montblanc/freeport/src/main.rs
@@ -1,7 +1,32 @@
 // Copyright © 2024 Stephan Kunz
 
+//! The node 'freeport' publishes an Int64 value every 50 ms on the topic /ganges
+//!
 //! This source is part of `DiMAS` implementation of Montblanc benchmark for distributed systems
 
-fn main() {
-	println!("Hello, world!");
+use dimas::prelude::*;
+use std::time::Duration;
+
+#[derive(Debug)]
+struct AgentProps {}
+
+#[tokio::main]
+async fn main() -> Result<()> {
+	let properties = AgentProps {};
+	let mut agent = Agent::new(Config::default(), properties);
+
+	agent
+		.timer()
+		.interval(Duration::from_millis(50))
+		.callback(|ctx, _props| {
+			let message = messages::Int64::random();
+			let value = message.data;
+			let _ = ctx.publish("ganges", message);
+			// just to see what value has been sent
+			println!("freeport sent: {value:>20}");
+		})
+		.add()?;
+
+	agent.start().await;
+	Ok(())
 }