+++ dimas/src/com/communicator.rs
@@ -15,30 +15,53 @@ use zenoh::publication::Publisher;
 // region:		--- Communicator
 #[derive(Debug)]
 pub struct Communicator {
-	// prefix to separate agents communicaton
-	pub(crate) prefix: String,
-	// the zenoh session
+	/// the zenoh session
 	pub(crate) session: Arc<Session>,
+	/// prefix to separate agents communicaton
+	pub(crate) prefix: Option<String>,
 }
 
 impl Communicator {
-	pub(crate) fn new(config: crate::config::Config, prefix: impl Into<String>) -> Self {
+	pub(crate) fn new(config: crate::config::Config) -> Self {
 		let cfg = config;
 		let session = Arc::new(
 			zenoh::open(cfg.zenoh_config())
 				.res_sync()
 				.expect("could not create zenoh session"),
 		);
-		let prefix = prefix.into();
-		Self { prefix, session }
+		Self {
+			session,
+			prefix: None,
+		}
+	}
+
+	pub(crate) fn new_with_prefix(
+		config: crate::config::Config,
+		prefix: impl Into<String>,
+	) -> Self {
+		let cfg = config;
+		let session = Arc::new(
+			zenoh::open(cfg.zenoh_config())
+				.res_sync()
+				.expect("could not create zenoh session"),
+		);
+		let prefix = Some(prefix.into());
+		Self { session, prefix }
 	}
 
 	pub(crate) fn uuid(&self) -> String {
 		self.session.zid().to_string()
 	}
 
-	pub(crate) fn prefix(&self) -> String {
+	pub(crate) fn prefix(&self) -> Option<String> {
 		self.prefix.clone()
+	}
+
+	pub(crate) fn key_expr(&self, msg_name: impl Into<String>) -> String {
+		match self.prefix.clone() {
+			Some(prefix) => prefix + "/" + &msg_name.into(),
+			None => msg_name.into(),
+		}
 	}
 
 	//#[cfg_attr(doc, doc(cfg(feature = "liveliness")))]
@@ -48,7 +71,7 @@ impl Communicator {
 		msg_type: impl Into<String> + Send,
 	) -> LivelinessToken<'a> {
 		let session = self.session.clone();
-		let uuid = self.prefix.clone() + "/" + &msg_type.into() + "/" + &session.zid().to_string();
+		let uuid = self.key_expr(msg_type) + "/" + &session.zid().to_string();
 		//dbg!(&uuid);
 		session
 			.liveliness()
@@ -74,7 +97,7 @@ impl Communicator {
 		T: bitcode::Encode,
 	{
 		let value: Vec<u8> = bitcode::encode(&message).expect("should never happen");
-		let key_expr = self.prefix.clone() + "/" + &msg_name.into();
+		let key_expr = self.key_expr(msg_name);
 		//dbg!(&key_expr);
 		match self.session.put(&key_expr, value).res_sync() {
 			Ok(()) => Ok(()),
@@ -83,7 +106,7 @@ impl Communicator {
 	}
 
 	pub(crate) fn delete(&self, msg_name: impl Into<String>) -> Result<()> {
-		let key_expr = self.prefix.clone() + "/" + &msg_name.into();
+		let key_expr = self.key_expr(msg_name);
 		//dbg!(&key_expr);
 		match self.session.delete(&key_expr).res_sync() {
 			Ok(()) => Ok(()),
@@ -101,7 +124,7 @@ impl Communicator {
 	) where
 		P: Send + Sync + Unpin + 'static,
 	{
-		let key_expr = self.prefix.clone() + "/" + &query_name.into();
+		let key_expr = self.key_expr(query_name);
 		//dbg!(&key_expr);
 		let ctx = ctx;
 		let props = props;
@@ -145,7 +168,7 @@ impl Communicator {
 
 impl Default for Communicator {
 	fn default() -> Self {
-		Self::new(crate::config::Config::local(), "peer")
+		Self::new(crate::config::Config::local())
 	}
 }
 // endregion:	--- Communicator
@@ -167,7 +190,7 @@ mod tests {
 	//#[serial]
 	async fn communicator_create_default() {
 		let _peer1 = Communicator::default();
-		let _peer2 = Communicator::new(crate::config::Config::local(), "peer2");
+		let _peer2 = Communicator::new_with_prefix(crate::config::Config::local(), "peer2");
 		//let _peer3 = Communicator::new(config::client());
 	}
 
@@ -175,20 +198,20 @@ mod tests {
 	//#[serial]
 	async fn communicator_create_single() {
 		let _peer1 = Communicator::default();
-		let _peer2 = Communicator::new(crate::config::Config::local(), "peer2");
+		let _peer2 = Communicator::new_with_prefix(crate::config::Config::local(), "peer2");
 	}
 
 	#[tokio::test(flavor = "multi_thread", worker_threads = 1)]
 	//#[serial]
 	async fn communicator_create_restricted() {
 		let _peer1 = Communicator::default();
-		let _peer2 = Communicator::new(crate::config::Config::local(), "peer2");
+		let _peer2 = Communicator::new_with_prefix(crate::config::Config::local(), "peer2");
 	}
 
 	#[tokio::test(flavor = "multi_thread")]
 	//#[serial]
 	async fn communicator_create_multi() {
 		let _peer1 = Communicator::default();
-		let _peer2 = Communicator::new(crate::config::Config::local(), "peer2");
+		let _peer2 = Communicator::new_with_prefix(crate::config::Config::local(), "peer2");
 	}
 }